+++ src/common/js/constants.js
@@ -22,5 +22,11 @@ EnvSettings.C = {
     // whether to wait for the Meteor startup() to load the configuration private assets
     // this should be done by the application, but the package is initialized before the former
     // starting with 2.0.0, we do not wait for startup, but is easy here to fallback to the historic behavior
-    WaitForStartup: false
+    WaitForStartup: false,
+
+    // an internal publication to let the server propagates the 'ready' status to the clients
+    _publicationName: 'EnvSettings.Ready',
+
+    // the collection managed by the internal publication above
+    _collectionName: 'EnvSettings.Ready'
 };